+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] - 2026-08-22
+
+### Changed
+- Renamed `MiniMapMarker` to `MiniMapMarkerUI` and updated the marker APIs exposed by `MiniMapUI`.
+- Added the custom editor icon metadata to `MiniMapMarkerUI`.
+
 ## [2.0.0] - 2026-08-22
 
 ### Added